+++ src/setup/index.spec.js
@@ -3,10 +3,10 @@ const validations = require('../utils/validations');
 const { names, warnings } = require('./constants');
 const questions = require('./questions');
 const tasks = require('./tasks');
+const { versionTransformer } = require('../utils');
 const {
   gitBranches,
   validateVersion,
-  versionTransformer,
   searchForIssues,
   askSetupQuestions,
   run
@@ -70,26 +70,6 @@ describe('validateVersion', () => {
       expect(validateVersion('a.b.c')).toBe(
         'The specified version is not a valid semver. Examples of valid semver: 1.2.3, 42.6.7.9.3-alpha, etc.'
       );
-    });
-  });
-});
-
-describe('versionTransformer', () => {
-  describe('when isFinal flag is true', () => {
-    const isFinal = true;
-
-    it('returns the version with the "v" prefix', () => {
-      expect(versionTransformer('9.11.1', [], { isFinal })).toBe('v9.11.1');
-      expect(versionTransformer('v9.11.1', [], { isFinal })).toBe('v9.11.1');
-    });
-  });
-
-  describe('when isFinal flag is false', () => {
-    const isFinal = false;
-
-    it('returns the version as input by the user', () => {
-      expect(versionTransformer('9.11.1', [], { isFinal })).toBe('9.11.1');
-      expect(versionTransformer('v9.11.1', [], { isFinal })).toBe('v9.11.1');
     });
   });
 });
